+++ ApiVuelos/src/main/java/com/ApiVuelos/ApiVuelos/controller/RouteController.java
@@ -5,6 +5,8 @@ import com.ApiVuelos.ApiVuelos.service.RouteService;
 import com.utn.tssi.tp5.Models.model.Airport;
 import com.utn.tssi.tp5.Models.model.Route;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.persistence.PersistenceException;
@@ -19,46 +21,66 @@ public class RouteController {
     @Autowired
     private RouteService routeService;
 
-    @PostMapping(value = "/add")
-    public void add(Airport begin ,Airport end,int distance,int time_estimed) {
+    @PostMapping(value = "/")
+    public ResponseEntity add(Airport begin , Airport end, int distance, int time_estimed) {
         try{
-            Route rt=new Route(begin,end,distance,time_estimed);
-            this.routeService.newObject(rt);
+            if(begin!=null && end!=null && distance!=0 && time_estimed!=0){
+                Route rt=new Route(begin,end,distance,time_estimed);
+                this.routeService.newObject(rt);
+                return new ResponseEntity(HttpStatus.OK);
+            }else{
+                return new ResponseEntity(HttpStatus.NO_CONTENT);
+            }
         }
-        catch(PersistenceException e){
-            e.printStackTrace();
+        catch(Exception e){
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @PutMapping(value = "/update")
-    public void update(Route value){
+    @PutMapping(value = "/")
+    public ResponseEntity update(Route value){
         try{
-            this.routeService.updateObject(value);
+            if(value!=null){
+                this.routeService.updateObject(value);
+                return new ResponseEntity(HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity(HttpStatus.NO_CONTENT);
+            }
         }
-        catch(PersistenceException e){
-            e.printStackTrace();
+        catch(Exception e){
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-
     }
 
-    @DeleteMapping(value = "/remove")
-    public void remove(Long id){
+    @DeleteMapping(value = "/")
+    public ResponseEntity remove(Long id){
         try{
-            this.routeService.removeObject(id);
-        }catch(PersistenceException e){
-            e.printStackTrace();
-        }
+            if(id!=null){
+                this.routeService.removeObject(id);
+                return new ResponseEntity(HttpStatus.OK);
+            }else
+            {
+                return new ResponseEntity(HttpStatus.NO_CONTENT);
+            }
 
+        }catch(Exception e){
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping(value = "/")
-    public List<Route> getAll() {
+    public ResponseEntity<List<Route>> getAll() {
         List<Route>rtList=new ArrayList<Route>();
         try{
             rtList=this.routeService.getAll();
-        }catch(PersistenceException e){
-            e.printStackTrace();
+            if(rtList.isEmpty()){
+                return new ResponseEntity<List<Route>>(HttpStatus.NO_CONTENT);
+            }else{
+                return new ResponseEntity<List<Route>>(rtList,HttpStatus.OK);
+            }
+        }catch(Exception e){
+            return new ResponseEntity<List<Route>>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return  rtList;
     }
 }
