+++ ApiVuelos/src/main/java/com/ApiVuelos/ApiVuelos/controller/StateController.java
@@ -5,6 +5,8 @@ import com.ApiVuelos.ApiVuelos.service.StateService;
 import com.utn.tssi.tp5.Models.model.Country;
 import com.utn.tssi.tp5.Models.model.State;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -23,58 +25,77 @@ public class StateController {
     @Autowired
     private CountryService countryService;
 
-    @RequestMapping(value = "/index.html")
-    public ModelAndView indexView() {
-        return null;
-    }
-
-    @PostMapping(value = "/add")
-    public void add(String name, String iataCode,String iso) {
+    @PostMapping(value = "/")
+    public ResponseEntity add(String name, String iataCode, String iso) {
         try{
-            Country country = this.countryService.getByAttributeType(iso);
+            if(name!=null && iataCode!=null && iso!=null){
+                Country country = this.countryService.getByAttributeType(iso);
 
-            if(country != null) {
-                State state = new State(name, iataCode, country);
-                this.stateService.newObject(state);
-
+                if(country != null) {
+                    State state = new State(name, iataCode, country);
+                    this.stateService.newObject(state);
+                    return  new ResponseEntity(HttpStatus.OK);
+                }
+                else
+                {
+                    return new ResponseEntity(HttpStatus.NO_CONTENT);
+                }
+            }else
+            {
+                return new ResponseEntity(HttpStatus.NO_CONTENT);
             }
         }
-        catch(PersistenceException e){
-            e.printStackTrace();
+        catch(Exception e){
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-
     }
-    @PutMapping(value = "/update")
-    public void update(State st) {
+    @PutMapping(value = "/")
+    public ResponseEntity update(State st) {
       try{
-            this.stateService.updateObject(st);
+          if(st!=null){
+              this.stateService.updateObject(st);
+              return new ResponseEntity(HttpStatus.OK);
+          }else
+          {
+              return new ResponseEntity(HttpStatus.NO_CONTENT);
+          }
       }
-      catch(PersistenceException e ){
-          e.printStackTrace();
+      catch(Exception e ){
+          return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
       }
     }
 
-    @DeleteMapping(value = "/remove")
-    public void remove(@RequestParam("id")Long id){
+    @DeleteMapping(value = "/")
+    public ResponseEntity remove(@RequestParam("id")Long id){
         try {
-            this.stateService.removeObject(id);
+            if(id!=null){
+                this.stateService.removeObject(id);
+                return new ResponseEntity(HttpStatus.OK);
+            }
+            else
+            {
+                return new ResponseEntity(HttpStatus.NO_CONTENT);
+            }
         }
-        catch(PersistenceException e){
-            e.printStackTrace();
+        catch(Exception e){
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-
     }
 
     @GetMapping(value ="/")
-    public List<State> getAll() {
+    public ResponseEntity<List<State>> getAll() {
         List<State>st=new ArrayList<State>();
                 try{
                     st= this.stateService.getAll();
+                    if(st.isEmpty()){
+                        return new ResponseEntity<List<State>>(HttpStatus.NO_CONTENT);
+                    }else
+                    {
+                        return new ResponseEntity<List<State>>(st,HttpStatus.OK);
+                    }
                 }
-                catch(PersistenceException e){
-                    e.printStackTrace();
+                catch(Exception e){
+                    return new ResponseEntity<List<State>>(HttpStatus.INTERNAL_SERVER_ERROR);
                 }
-
-        return st;
     }
 }
